# backend/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestão
  description: API para gerenciar dedicações, funcionários, projetos e períodos de férias
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
paths:
  /projects-with-details:
    get:
      summary: Retrieve all projects with groups and vacations
      description: Returns a list of all projects, where each project contains a list of dedicated workers (groups) and their vacation periods (items).
      tags:
        - Projects
      responses:
        200:
          description: A list of projects with detailed groups and vacation items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectWithDetails'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dedications-with-details:
    get:
      summary: Retrieve all dedications with groups and items
      description: Returns a list of all dedications, organized into items (dedications) and groups (workers with dedications).
      tags:
        - Dedications
      responses:
        200:
          description: A list of dedications with detailed groups and items.
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    description: A list of workers with dedications
                    items:
                      $ref: '#/components/schemas/Group'
                  items:
                    type: array
                    description: A list of dedications for the workers in the projects
                    items:
                      $ref: '#/components/schemas/DedicationItem'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dedications:
    post:
      summary: Criar uma nova dedicação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dedication'
      responses:
        '201':
          description: Dedicação criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dedication'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Listar todas as dedicações
      responses:
        '200':
          description: Lista de dedicações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dedication'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dedications/{id}:
    get:
      summary: Buscar uma dedicação específica por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID da dedicação
      responses:
        '200':
          description: Dedicação encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dedication'
        '404':
          description: Dedicação não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Atualizar uma dedicação
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID da dedicação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dedication'
      responses:
        '200':
          description: Dedicação atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dedication'
        '404':
          description: Dedicação não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deletar uma dedicação
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID da dedicação
      responses:
        '204':
          description: Dedicação deletada com sucesso
        '404':
          description: Dedicação não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees:
    post:
      summary: Criar um novo funcionário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Funcionário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Listar todos os funcionários
      responses:
        '200':
          description: Lista de funcionários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /employees/{id}:
    get:
      summary: Buscar um funcionário por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do funcionário
      responses:
        '200':
          description: Funcionário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Funcionário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Atualizar um funcionário
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do funcionário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Funcionário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '404':
          description: Funcionário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deletar um funcionário
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do funcionário
      responses:
        '200':
          description: Funcionário deletado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Funcionário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects:
    post:
      summary: Criar um novo projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Projeto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Listar todos os projetos
      responses:
        '200':
          description: Lista de projetos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{id}:
    get:
      summary: Buscar um projeto por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do projeto
      responses:
        '200':
          description: Projeto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Atualizar um projeto
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do projeto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Projeto atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deletar um projeto
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do projeto
      responses:
        '200':
          description: Projeto deletado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Projeto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /vacations:
    post:
      summary: Criar um novo período de férias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vacation'
      responses:
        '201':
          description: Período de férias criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vacation'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Listar todos os períodos de férias
      responses:
        '200':
          description: Lista de períodos de férias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vacation'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /vacations/{id}:
    get:
      summary: Buscar um período de férias por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do período de férias
      responses:
        '200':
          description: Período de férias encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vacation'
        '404':
          description: Período de férias não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Atualizar um período de férias
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do período de férias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vacation'
      responses:
        '200':
          description: Período de férias atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vacation'
        '404':
          description: Período de férias não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Deletar um período de férias
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID do período de férias
      responses:
        '204':
          description: Período de férias deletado com sucesso
        '404':
          description: Período de férias não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:


    DedicationItem:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the dedication within the request
          example: 9
        content:
          type: string
          description: Project name and percentage dedication
          example: Project Name - 30%
        start:
          type: string
          format: date
          description: Start date of the dedication
          example: 2025-11-08
        end:
          type: string
          format: date
          description: End date of the dedication
          example: 2025-11-14
        group:
          type: integer
          description: The ID of the group (worker) associated with this dedication
          example: 3
    ProjectWithDetails:
        type: object
        properties:
          project_name:
            type: string
            description: The name of the project
            example: Project A
          groups:
            type: array
            description: A list of workers dedicated to the project
            items:
              $ref: '#/components/schemas/Group'
          items:
            type: array
            description: A list of vacation periods for the workers in the project
            items:
              $ref: '#/components/schemas/VacationItem'
    Group:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the group within the request
          example: 1
        content:
          type: string
          description: Name of the worker
          example: Bruno
    VacationItem:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the vacation item within the request
          example: vacation-0-0
        content:
          type: string
          description: The duration of the vacation in days
          example: 6 days
        start:
          type: string
          format: date
          description: Start date of the vacation
          example: 2025-11-08
        end:
          type: string
          format: date
          description: End date of the vacation
          example: 2025-11-14
        group:
          type: integer
          description: The ID of the group (worker) associated with this vacation
          example: 1
        className:
          type: string
          description: The CSS class name for the vacation, used in UI rendering
          example: designer task

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A description of the error
          example: Internal server error
    Dedication:
      type: object
      properties:
        employee_name:
          type: string
        project:
          type: string
        initial_date_real:
          type: string
          format: date
        final_date_real:
          type: string
          format: date
        percentage:
          type: number
          format: float
    Employee:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        projects:
          type: array
          items:
            type: object
            properties:
              project:
                type: string
              percentage:
                type: number
                format: float
        vacations:
          type: array
          items:
            type: object
            properties:
              start_date:
                type: string
                format: date
              end_date:
                type: string
                format: date
    Project:
      type: object
      properties:
        project_name:
          type: string
        initial_date_estimated:
          type: string
          format: date
        final_date_estimated:
          type: string
          format: date
        initial_date_real:
          type: string
          format: date
        final_date_real:
          type: string
          format: date
    Vacation:
      type: object
      properties:
        employee:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
    Error:
      type: object
      properties:
        error:
          type: string
    Message:
      type: object
      properties:
        message:
          type: string